requirements.txt
pydub
SpeechRecognition
openai-whisper
torch
numpy
openai
click
python-dotenv
pyaudio
aiofiles 